requirements.txt
numpy==1.20.1
pygame==2.1.0
PyOpenGL==3.1.5
